requirements.txt
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
Pillow
numpy
scikit-learn
pandas
joblib